--- app/src/main/java/ru/beryukhov/coffeegram/ListExt.kt
@@ -1,9 +1,0 @@
-package ru.beryukhov.coffeegram
-
-internal operator fun <E> List<E>.times(i: Int): List<E> {
-    val result = mutableListOf<E>()
-    (0 until i).forEach {
-        result.addAll(this)
-    }
-    return result
-}
